% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_bathy_diff.R
\name{calc_bathy_diff}
\alias{calc_bathy_diff}
\title{Calculate the difference between observed and predicted bathymetry}
\usage{
calc_bathy_diff(obs, pred)
}
\arguments{
\item{obs}{SpatRaster object with the observed bathymetry.}

\item{pred}{SpatRaster object with the predicted bathymetry.}
}
\value{
SpatRaster object with the difference between the observed and
predicted bathymetry.
}
\description{
Calculate the difference between observed and predicted bathymetry
}
