% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_lake_volume.R
\name{calculate_lake_volume}
\alias{calculate_lake_volume}
\title{Calculate the volume of a lake using bathymetry data or a hypsograph}
\usage{
calculate_lake_volume(
  bathy_raster = NULL,
  hyps = NULL,
  depth = NULL,
  return_rast = FALSE
)
}
\arguments{
\item{bathy_raster}{SpatRaster object with the bathymetry data.}

\item{hyps}{data.frame with columns 'depth' and 'area'.}

\item{depth}{numeric. The depth to which to calculate the volume. If
provided, the volume will be calculated to this depth. If not provided, the
volume will be calculated to the maximum depth of the bathymetry raster or
the hypsograph.}

\item{return_rast}{logical. If TRUE, return a raster with the calculated
volume in each grid cell. Default is FALSE.}
}
\value{
numeric. The volume of the lake in cubic meters (m^3).
If \code{return_rast} is TRUE, a SpatRaster object with the volume
in each grid cell is returned.
}
\description{
Calculate the volume of a lake using bathymetry data or a hypsograph
}
\examples{
shoreline <- readRDS(system.file("extdata/rotoma_shoreline.rds",
package = "bathytools"))
depth_points <- readRDS(system.file("extdata/depth_points.rds",
package = "bathytools"))
bathy_raster <- rasterise_bathy(shoreline = shoreline,
depth_points = depth_points, crs = 2193)
calculate_lake_volume(bathy_raster = bathy_raster)

}
