% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_for_islands.R
\name{check_for_islands}
\alias{check_for_islands}
\title{Check for islands in a set of polygons}
\usage{
check_for_islands(x)
}
\arguments{
\item{x}{sf POLYGON object}
}
\value{
sf POLYGON object with an additional column \code{is_island} indicating
whether the polygon is an island or not.
}
\description{
Check for islands in a set of polygons
}
