% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_depth_points.R
\name{generate_depth_points}
\alias{generate_depth_points}
\title{Generate depth points for interpolation}
\usage{
generate_depth_points(
  shoreline,
  islands = NULL,
  depth_points = NULL,
  contours = NULL,
  res = 2,
  subsample = TRUE,
  crs = NULL
)
}
\arguments{
\item{shoreline}{sf object of lake shoreline.}

\item{islands}{sf object of lake islands if present.}

\item{depth_points}{sf object of depth points or a dataframe of points with
columns 'lon' and 'lat'. Must contain a "depth" column. If NULL, then
contours must be provided.
If NULL, then contours must be provided.}

\item{contours}{sf object of depth contours. If NULL, then points must be
provided.}

\item{res}{numeric resolution of output raster in metres.}

\item{subsample}{logical subsample points to 10000 if TRUE. Default is TRUE.}

\item{crs}{target coordinate reference system: object of class \code{crs}, or input string for \link[sf]{st_crs}}
}
\value{
sf object of depth points.
}
\description{
Generate depth points for interpolation
}
\examples{
shoreline <- readRDS(system.file("extdata/rotoma_shoreline.rds",
package = "bathytools"))
depth_points <- readRDS(system.file("extdata/depth_points.rds",
package = "bathytools"))

# Generate depth points with points ----
depth_points <- generate_depth_points(shoreline = shoreline,
depth_points = depth_points)

# Generate depth points with contours ----
contours <- readRDS(system.file("extdata/depth_contours.rds",
package = "bathytools"))
depth_points <- generate_depth_points(shoreline = shoreline,
contours = contours)

}
