% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_contours.R
\name{get_contours}
\alias{get_contours}
\title{Get contours from a bathymetric raster}
\usage{
get_contours(bathy_raster, surface = 0, depths = 1)
}
\arguments{
\item{bathy_raster}{SpatRaster object with bathymetry depths (negative values).}

\item{surface}{numeric. The surface elevation of the lake. Default is 0.}

\item{depths}{numeric. The depths at which to calculate the area. If a single
numeric value is provided, the function will calculate the area at each depth
from the surface to the minimum depth of the bathymetry raster at intervals of
the provided value. If a vector of numeric values is provided, the function
will calculate the area at each depth specified in the vector. Default is 1.}
}
\value{
sf LINESTRING of the contours.
}
\description{
Get contours from a bathymetric raster
}
\examples{
shoreline <- readRDS(system.file("extdata/rotoma_shoreline.rds",
package = "bathytools"))
point_data <- readRDS(system.file("extdata/depth_points.rds",
package = "bathytools"))
bathy_raster <- rasterise_bathy(shoreline = shoreline,
point_data = point_data, crs = 2193, res = 8)
contours <- get_contours(bathy_raster = bathy_raster)
}
