% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lake_depth.R
\name{get_lake_depth}
\alias{get_lake_depth}
\title{Get Lake Depth from Bathymetry Raster
This function calculates the lake depth from a bathymetric raster.}
\usage{
get_lake_depth(bathy_raster)
}
\arguments{
\item{bathy_raster}{SpatRaster object with the bathymetry data.}
}
\value{
numeric. The lake depth.
}
\description{
Get Lake Depth from Bathymetry Raster
This function calculates the lake depth from a bathymetric raster.
}
