% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lake_surface_elevation.R
\name{get_lake_surface_elevation}
\alias{get_lake_surface_elevation}
\title{Get lake surface elevation from DEM}
\usage{
get_lake_surface_elevation(dem_raster, shoreline)
}
\arguments{
\item{dem_raster}{SpatRaster object with the DEM data.}

\item{shoreline}{sf object of lake shoreline.}
}
\value{
numeric
}
\description{
Get lake surface elevation from DEM
}
\examples{
shoreline <- readRDS(system.file("extdata/rotoma_shoreline.rds",
package = "bathytools"))
dem_raster <- terra::rast(system.file("extdata/dem_32m.tif",
package = "bathytools"))
lake_elev <- get_lake_surface_elevation(dem_raster = dem_raster,
shoreline = shoreline)
}
