% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_shoreline.R
\name{get_shoreline}
\alias{get_shoreline}
\title{Get the shoreline from a bathymetric raster}
\usage{
get_shoreline(bathy_raster, dTolerance = NULL)
}
\arguments{
\item{bathy_raster}{SpatRaster object with bathymetry depths (negative values).}

\item{dTolerance}{numeric. The tolerance for simplifying the shoreline.
Default is the maximum resolution of the bathymetric raster. If set to 0, the
shoreline will not be simplified.}
}
\value{
sf POLYGON of the shoreline.
}
\description{
Get the shoreline from a bathymetric raster
}
\examples{
shoreline <- readRDS(system.file("extdata/rotoma_shoreline.rds",
package = "bathytools"))
depth_points <- readRDS(system.file("extdata/depth_points.rds",
package = "bathytools"))
bathy_raster <- rasterise_bathy(shoreline = shoreline,
depth_points = depth_points, crs = 2193, res = 8)
shoreline2 <- get_shoreline(bathy_raster = bathy_raster)
}
