% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_bathy_dem.R
\name{merge_bathy_dem}
\alias{merge_bathy_dem}
\title{Merge lake bathymetry raster with a DEM raster}
\usage{
merge_bathy_dem(
  shoreline,
  bathy_raster,
  dem_raster,
  transform_shoreline = TRUE,
  catchment = NULL,
  crop_dem_to_catchment = TRUE,
  blend_buffer = 10
)

merge_bathy_dem(
  shoreline,
  bathy_raster,
  dem_raster,
  transform_shoreline = TRUE,
  catchment = NULL,
  crop_dem_to_catchment = TRUE,
  blend_buffer = 10
)
}
\arguments{
\item{shoreline}{sf object of lake shoreline.}

\item{bathy_raster}{SpatRaster object with bathymetry depths (negative values).}

\item{dem_raster}{SpatRaster object with DEM elevations.}

\item{transform_shoreline}{logical. If TRUE, reprojects shoreline to match DEM CRS.}

\item{catchment}{sf object of the catchment boundary.}

\item{crop_dem_to_catchment}{logical. If TRUE, crops DEM to catchment boundary.}

\item{blend_buffer}{numeric. Width (map units) of the shoreline blending zone.}
}
\value{
SpatRaster object with the merged bathymetry and DEM data.

SpatRaster object with merged bathymetry and DEM data.
}
\description{
Merge lake bathymetry raster with a DEM raster

Merge lake bathymetry raster with a DEM raster (smooth blend at shoreline)
}
\examples{
shoreline <- readRDS(system.file("extdata/rotoma_shoreline.rds",
package = "bathytools"))
catchment <- readRDS(system.file("extdata/rotoma_catchment.rds",
                                 package = "bathytools"))
depth_points <- readRDS(system.file("extdata/depth_points.rds",
package = "bathytools"))
bathy_raster <- rasterise_bathy(shoreline = shoreline,
depth_points = depth_points, crs = 2193)
dem_raster <- terra::rast(system.file("extdata/dem_32m.tif",
package = "bathytools"))
dem_bath <- merge_bathy_dem(shoreline = shoreline, bathy_raster = bathy_raster,
dem_raster = dem_raster, catchment = catchment)
}
