% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_raster_3d.R
\name{plot_raster_3d}
\alias{plot_raster_3d}
\title{Plot a 3D raster}
\usage{
plot_raster_3d(
  x,
  fact = 0,
  cols = "topo",
  split_lake = FALSE,
  mask_to_lake = FALSE,
  shoreline = NULL,
  lake_elev = NULL,
  add_shoreline = FALSE,
  aspectratio_z = 0.2,
  dim_px = 300
)
}
\arguments{
\item{x}{SpatRaster object}

\item{fact}{positive integer. Aggregation factor expressed as number of cells in each direction (horizontally and vertically). Or two integers (horizontal and vertical aggregation factor) or three integers (when also aggregating over layers)}

\item{mask_to_lake}{logical, mask the raster to the lake. Default is FALSE.}

\item{shoreline}{sf object of lake shoreline.}

\item{lake_elev}{numeric, elevation of the lake surface. Can be extracted using
the \link{get_lake_surface_elevation} function.}

\item{aspectratio_z}{numeric, aspect ratio of the z axis. Default is 0.2.}

\item{dim_px}{numeric, dimension of the plot in pixels. Default is 300.}
}
\value{
plotly object
}
\description{
Plot a 3D raster
}
\examples{
r <- terra::rast(system.file("ex/elev.tif", package="terra"))
plot_raster_3d(r)
}
