% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasterise_bathy.R
\name{rasterise_bathy}
\alias{rasterise_bathy}
\title{Generate depth points for interpolation}
\usage{
rasterise_bathy(
  shoreline,
  islands = NULL,
  point_data = NULL,
  contours = NULL,
  res = 2,
  subsample = TRUE,
  crs,
  method = "MBA",
  print_plot = TRUE,
  n = 1,
  m = 1,
  h = 8
)
}
\arguments{
\item{shoreline}{sf object of lake shoreline.}

\item{islands}{sf object of lake islands if present.}

\item{point_data}{sf object of depth points. Must contain a "depth" column.
If NULL, then contours must be provided.}

\item{contours}{sf object of depth contours. If NULL, then points must be
provided.}

\item{res}{numeric resolution of output raster in metres.}

\item{subsample}{logical subsample points to 10000 if TRUE. Default is TRUE.}

\item{crs}{target coordinate reference system: object of class \code{crs}, or input string for \link[sf]{st_crs}}

\item{method}{character interpolation method. Options are 'MBA' (default),}

\item{print_plot}{logical print plot of interpolated raster.}

\item{n}{initial size of the spline space in the hierarchical
    construction along the x axis. If the rectangular domain is a
    square, n = m = 1 is recommended. If the x axis is k times the length
    of the y axis, n = 1, m = k is recommended. The default is n = 1. }

\item{m}{initial size of the spline space in the hierarchical
    construction along the y axis. If the y axis is k times the length
    of the x axis, m = 1, n = k is recommended. The default is m = 1.  }

\item{h}{Number of levels in the hierarchical construction. If, e.g.,
    n = m = 1 and h = 8, the resulting spline surface has a coefficient
    grid of size \eqn{2^h}{2^h} + 3 = 259 in each direction of the
    spline surface. See references for additional information. }
}
\value{
sf object of depth points.
}
\description{
Generate depth points for interpolation
}
\examples{
shoreline <- readRDS(system.file("extdata/rotoma_shoreline.rds",
package = "bathytools"))
point_data <- readRDS(system.file("extdata/depth_points.rds",
package = "bathytools"))
bathy <- rasterise_bathy(shoreline = shoreline, point_data = point_data,
crs = 2193)

}
