% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_dem_bath.R
\name{tm_dem_bath}
\alias{tm_dem_bath}
\title{Draw a elevation map with bathymetry and surrounding topography}
\usage{
tm_dem_bath(dem_bath, lake_elev, ext_elev = 0)
}
\arguments{
\item{dem_bath}{SpatRaster object with merged DEM and bathymetry data. Can be
generated using the \link{merge_bathy_dem} function.}

\item{lake_elev}{numeric, elevation of the lake surface. Can be extracted using
the \link{get_lake_surface_elevation} function.}

\item{ext_elev}{numeric, additional elevation to extend the lake surface.
This allows you to generate a flood map. Default is 0.}
}
\description{
Draw a elevation map with bathymetry and surrounding topography
}
