% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_lake_volume.R
\name{calculate_lake_volume}
\alias{calculate_lake_volume}
\title{Calculate the volume of a lake using bathymetry data or a hypsograph}
\usage{
calculate_lake_volume(bathy_raster = NULL, hyps = NULL, depth = NULL)
}
\arguments{
\item{bathy_raster}{SpatRaster object with the bathymetry data.}

\item{hyps}{data.frame with columns 'depth' and 'area'.}

\item{depth}{numeric. The depth to which to calculate the volume. If
provided, the volume will be calculated to this depth. If not provided, the
volume will be calculated to the maximum depth of the bathymetry raster or
the hypsograph.}
}
\value{
numeric. The volume of the lake in cubic meters.
}
\description{
Calculate the volume of a lake using bathymetry data or a hypsograph
}
\examples{
shoreline <- readRDS(system.file("extdata/rotoma_shoreline.rds",
package = "bathytools"))
point_data <- readRDS(system.file("extdata/depth_points.rds",
package = "bathytools"))
bathy_raster <- rasterise_bathy(shoreline = shoreline,
point_data = point_data, crs = 2193)
calculate_lake_volume(bathy_raster = bathy_raster)

}
