% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_bathy_dem.R
\name{merge_bathy_dem}
\alias{merge_bathy_dem}
\title{Merge lake bathymetry raster with a DEM raster}
\usage{
merge_bathy_dem(
  shoreline,
  bathy_raster,
  dem_raster,
  transform_shoreline = TRUE,
  catchment = NULL,
  crop_dem_to_catchment = TRUE
)
}
\arguments{
\item{shoreline}{sf object of lake shoreline.}

\item{bathy_raster}{SpatRaster object with the bathymetry data.}

\item{dem_raster}{SpatRaster object with the DEM data.}

\item{transform_shoreline}{logical. If TRUE, the shoreline will be
transformed to match the CRS of the DEM raster.}

\item{catchment}{sf object of the catchment boundary. If provided, the DEM
raster will be cropped to the catchment boundary.}

\item{crop_dem_to_catchment}{logical. If TRUE, the DEM raster will be cropped
to the catchment boundary. Requires catchment to be provided.}
}
\value{
SpatRaster object with the merged bathymetry and DEM data.
}
\description{
Merge lake bathymetry raster with a DEM raster
}
\examples{
shoreline <- readRDS(system.file("extdata/rotoma_shoreline.rds",
package = "bathytools"))
catchment <- readRDS(system.file("extdata/rotoma_catchment.rds",
                                 package = "bathytools"))
point_data <- readRDS(system.file("extdata/depth_points.rds",
package = "bathytools"))
bathy_raster <- rasterise_bathy(shoreline = shoreline,
point_data = point_data, crs = 2193)
dem_raster <- terra::rast(system.file("extdata/dem_32m.tif",
package = "bathytools"))
dem_bath <- merge_bathy_dem(shoreline = shoreline, bathy_raster = bathy_raster,
dem_raster = dem_raster, catchment = catchment)
}
